% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Generate DEGify heatmap}
\usage{
plot_heatmap(
  vsd,
  res_df,
  metadata,
  top_n_heatmap = 50,
  padj_cutoff = 0.05,
  output_dir = "DEGify_Output"
)
}
\arguments{
\item{vsd}{Variance stabilized dataset.}

\item{res_df}{Differential expression results dataframe.}

\item{metadata}{Sample metadata dataframe.}

\item{top_n_heatmap}{Number of genes displayed
in the heatmap.}

\item{padj_cutoff}{Adjusted p-value threshold.}

\item{output_dir}{Directory for exported plots.}
}
\value{
Heatmap expression matrix.
}
\description{
Generates a clustered heatmap of top
differentially expressed genes.
}
