% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_pca}
\alias{plot_pca}
\title{Generate PCA plot}
\usage{
plot_pca(dds, metadata, output_dir = "DEGify_Output")
}
\arguments{
\item{dds}{DESeq2 dataset object.}

\item{metadata}{Sample metadata dataframe.}

\item{output_dir}{Directory for exported plots.}
}
\value{
A list containing:
\itemize{
\item Variance stabilized dataset
\item PCA ggplot object
}
}
\description{
Performs variance stabilizing transformation
and generates a PCA plot.
}
