% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Generate DEGify volcano plot}
\usage{
plot_volcano(res_df, top_n_labels = 10, output_dir = "DEGify_Output")
}
\arguments{
\item{res_df}{Differential expression results dataframe.}

\item{top_n_labels}{Number of top genes labeled
on the volcano plot.}

\item{output_dir}{Directory for exported plots.}
}
\value{
A ggplot2 volcano plot object.
}
\description{
Creates a volcano plot for differential
expression analysis results.
}
