% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichement.R
\name{run_go_enrichment}
\alias{run_go_enrichment}
\title{Run Gene Ontology enrichment analysis}
\usage{
run_go_enrichment(sig_deg, ontology = "BP", output_dir = "DEGify_Output")
}
\arguments{
\item{sig_deg}{Significant DEG dataframe.}

\item{ontology}{GO ontology category.
One of "BP", "MF", or "CC".}

\item{output_dir}{Directory for exported results.}
}
\value{
A list containing:
\itemize{
\item GO enrichment results
\item GO barplot
\item GO dotplot
}
}
\description{
Performs GO enrichment analysis on
significant differentially expressed genes.
}
